refactor(e2e): break e2e tests into atomic files
