+++ integration/jwtserver_test.go
@@ -1,0 +1,7 @@
+//go:build e2e
+
+package integration
+
+type E2EJWTServer struct {
+	IntegrationTestSuite
+}
